requirements.txt
flask
jellyfish
svg.charts